// shared/models/enum/filter-keys.enum.ts
export enum FilterKey {
  NONE = "",
  MEDIA_TYPE = "media-type",
  LAST_EDITED = "edited-dates",
  CREATED = "created",
  EPISODES_CURRENTLY_IN_TELEVISION = "episodes-currently-in-television",
  EPISODES_IN_TELEVISION = "episodes-in-television",
  EPISODE_DETAILS = "episode-details",
  URLS = "urls",
  TAGS = "tags",
  URLS_WATCH = "urls-watch",
  URLS_VIDEO = "urls-video",
  URLS_INFO = "urls-info",
  CHANNELS = "channels",
  CHANNEL_TYPES = "channel-types",
  LANGUAGES = "languages",
  COUNTRIES = "countries",
  SOURCES = "sources",
  RATING = "rating",
  RATING_IMDB = "rating-imdb",
  RATING_METASCORE = "rating-metascore",
  RATING_WATCHABILITY = "rating-watchability",
  STATUS = "status",
  GENRES = "genres",
  FAVORITE = "favorite",
  NOTE = "note",
  LINKED_MEDIA = "linked-media",
  LINKED_RECIPES = "linked-recipes",
  IMAGE = "image",
  LIVE = "live",
  TELEVISION = "television",
  DVD = "dvd",
  CINEMA = "cinema",
  YEAR_START = "year-start",
  YEAR_END = "year-end",
  TAGLINE = "tagline",
  AUTOMATIC = "automatic",
  EPISODE_NUMBERING = "episode-numbering",
  WRAP_EPISODES = "wrap-episodes",
  YEAR_PER_SEASON = "year-per-season",
  PREPARED = "prepared",
  PREPARED_DATES = "prepared-dates",
  PLANNED = "planned",
  PLANNED_DATES = "planned-dates",
  EDITED_DATES = "edited-dates",
  CATEGORIES = "categories",
  DIFFICULTIES = "difficulties",
  CONTENTS = "contents",
  BASIC_RECIPE = "basic-recipe",
  ON_SHOPPINGLIST = "on-shoppinglist",
  COSTS = "costs",
  INGREDIENTS_QUANTITY = "ingredient-quantity",
  PREPARATION_TIME = "preparation-time",
  PORTIONS_LEFT = "portions-left",
  RUNTIME = "runtime",
  AVAILABLE_UNTIL = "available-until",
  YEARS = "years",
  INTERNATIONAL = "international",
}